requirements.txt
black==22.3.0
click==8.0.3
colorama==0.4.4
hilbertcurve==2.0.5
mypy-extensions==0.4.3
numpy==1.22.2
opencv-python==4.5.5.64
pandas==1.4.0
pathspec==0.9.0
Pillow==9.0.1
platformdirs==2.5.0
python-dateutil==2.8.2
python-rle==0.0.3
pytz==2021.3
six==1.16.0
tomli==2.0.1
torch==1.11.0.dev20220203+cpu
torchvision==0.2.2.post3
tqdm==4.62.3
typing_extensions==4.0.1
websockets==10.1
yapf==0.32.0
